Moved String_Array into its own file.
